Add a virtual boundary-element
A virtual boundary is an element that wraps a DOM-tree and takes on
the dimension of the first child or an offset to that first child if
specified. These boundaries are useful when causing effects that are
relative to it's position and bounding-rectangle. The most typical
use-case is to pre-load content before executing an actual action like
a click-event.